Make implicit conversion explicit
This seems somewhat strange to me - double -> long double -> double again (as the value of floating_point_t), but it seems the conditional statement and the va_arg(long double) reference cause the compiler to implicitly promote the second statement.

+++ src/printf/printf.c
@@ -1297,7 +1297,7 @@ static inline void format_string_loop(output_gadget_t* output, const char* forma
 #if PRINTF_SUPPORT_DECIMAL_SPECIFIERS
       case 'f' :
       case 'F' : {
-        floating_point_t value = (floating_point_t) (flags & FLAGS_LONG_DOUBLE ? va_arg(args, long double) : va_arg(args, double));
+        floating_point_t value = (floating_point_t) (flags & FLAGS_LONG_DOUBLE ? va_arg(args, long double) : (long double) va_arg(args, double));
         if (*format == 'F') flags |= FLAGS_UPPERCASE;
         print_floating_point(output, value, precision, width, flags, PRINTF_PREFER_DECIMAL);
         format++;
@@ -1309,7 +1309,7 @@ static inline void format_string_loop(output_gadget_t* output, const char* forma
       case 'E':
       case 'g':
       case 'G': {
-        floating_point_t value = (floating_point_t) (flags & FLAGS_LONG_DOUBLE ? va_arg(args, long double) : va_arg(args, double));
+        floating_point_t value = (floating_point_t) (flags & FLAGS_LONG_DOUBLE ? va_arg(args, long double) : (long double) va_arg(args, double));
         if ((*format == 'g')||(*format == 'G')) flags |= FLAGS_ADAPT_EXP;
         if ((*format == 'E')||(*format == 'G')) flags |= FLAGS_UPPERCASE;
         print_floating_point(output, value, precision, width, flags, PRINTF_PREFER_EXPONENTIAL);
